Adding 'default\2_multi_year' problem (not working). Code refactoring required. Last commit before branching.
