+++ default/2_multi_year/sets_structure.yml
@@ -1,0 +1,73 @@
+# structure of sets
+
+scenarios:
+    symbol: s
+    table_name: _set_SCENARIOS
+    table_headers:
+        id: ['s_ID', 'INTEGER PRIMARY KEY']
+        name: ['s_Name', 'TEXT']
+    split_problem: True
+
+sensitivity:
+    symbol: e
+    table_name: _set_SENSITIVITY
+    table_headers:
+        id: ['e_ID', 'INTEGER PRIMARY KEY']
+        name: ['e_Name', 'TEXT']
+    split_problem: True
+
+time:
+    symbol: y
+    table_name: _set_TIME
+    table_headers:
+        id: ['y_ID', 'INTEGER PRIMARY KEY']
+        name: ['y_Name', 'TEXT']
+
+stock:
+    symbol: k
+    table_name: _set_STOCKS
+    table_headers:
+        id: ['k_ID', 'INTEGER PRIMARY KEY']
+        name: ['k_Name', 'TEXT']
+        category: ['k_Category', 'TEXT']
+    set_categories:
+        k.o: Operative capacity
+        k.n: New capacity
+
+technologies:
+    symbol: t
+    table_name: _set_TECHNOLOGIES
+    table_headers:
+        id: ['t_ID', 'INTEGER PRIMARY KEY']
+        name: ['t_Name', 'TEXT']
+        category: ['t_Category', 'TEXT']
+    set_categories:
+        t.s: Supply technology
+        t.d: Demand technology
+
+flows:
+    symbol: f
+    table_name: _set_FLOWS
+    table_headers:
+        id: ['f_ID', 'INTEGER PRIMARY KEY']
+        name: ['f_Name', 'TEXT']
+        category: ['f_Category', 'TEXT']
+        competition: ['f_Competition', 'TEXT']
+        parent: ['f_Parent_technology', 'TEXT']
+        unit: ['f_Unit', 'TEXT']
+    set_categories:
+        f.p: Product flow
+        f.co: Operational costs
+        f.ca: Capital costs
+
+technical_specs:
+    symbol: ts
+    table_name: _set_TECHNICAL_SPECS
+    table_headers:
+        id: ['ts_ID', 'INTEGER PRIMARY KEY']
+        name: ['ts_Name', 'TEXT']
+        category: ['ts_Category', 'TEXT']
+        unit: ['ts_Unit', 'TEXT']
+    set_categories:
+        ts.max: Availability, max
+        ts.min: Availability, min
