+++ default/2_multi_year/variables.yml
@@ -1,0 +1,225 @@
+# template for defining variables of the model
+
+# RULES:
+# - variables define SQLite tables (name case insensitive: two tables named 
+# 'e' and 'E' are not allowed. However, variables symbols are case sensitive).
+# - 
+
+
+# -----------------------------------------
+# exogenous variables
+
+I:
+    symbol: I
+    name: special identity matrix
+    type: constant
+    coordinates_info: 
+        flows: 
+            set_categories: f.p
+    shape: [1, flows]
+    value: sum_vector
+
+T: 
+    symbol: T
+    name: lower triangular matrix
+    type: constant
+    coordinates_info: 
+        time: null
+    shape: [time, time]
+    value: lower_triangular
+
+u:
+    symbol: u
+    name: use coefficients matrix
+    type: exogenous
+    coordinates_info:
+        scenarios: null
+        technologies: 
+            set_categories: t.s
+        flows:
+            set_categories: f.p
+            aggregation_key: competition
+    shape: [flows, technologies]
+
+d:
+    symbol: d
+    name: market shares coefficients matrix
+    type: exogenous
+    coordinates_info:
+        scenarios: null
+        technologies: 
+            set_categories: t.s
+        flows:
+            set_categories: f.p
+    shape: [technologies, flows]
+
+Y:
+    symbol: Y
+    name: total final demand
+    type: exogenous
+    coordinates_info:
+        scenarios: null
+        time: null
+        technologies: 
+            set_categories: t.d
+        flows: 
+            set_categories: f.p
+            aggregation_key: competition
+    shape: [flows, time]
+
+cap_oi:
+    symbol: cap_oi
+    name: operative capacity, initial
+    type: exogenous
+    coordinates_info:
+        scenarios: null
+        technologies:
+            set_categories: t.s
+    shape: [1, technologies]
+
+c_op_spec:
+    symbol: c_op
+    name: operational costs
+    type: exogenous
+    coordinates_info:
+        scenarios: null
+        sensitivity: null
+        technologies:
+            set_categories: t.s
+        flows:
+            set_categories: f.co
+    shape: [flows, technologies]
+
+c_ca_spec:
+    symbol: c_ca
+    name: capital costs
+    type: exogenous
+    coordinates_info:
+        scenarios: null
+        sensitivity: null
+        technologies:
+            set_categories: t.s
+        flows:
+            set_categories: f.ca
+    shape: [flows, technologies]
+
+ts_max:
+    symbol: ts_max
+    name: availability, max
+    type: exogenous
+    coordinates_info:
+        scenarios: null
+        technologies:
+            set_categories: t.s
+        technical_specs:
+            set_categories: ts.max
+    shape: [technical_specs, technologies]
+
+ts_min:
+    symbol: ts_min
+    name: availability, min
+    type: exogenous
+    coordinates_info:
+        scenarios: null
+        technologies:
+            set_categories: t.s
+        technical_specs:
+            set_categories: ts.min
+    shape: [technical_specs, technologies]
+
+
+# -----------------------------------------
+# endogenous variables
+
+Q:
+    symbol: Q
+    name: total demand by flow
+    type: endogenous
+    coordinates_info:
+        scenarios: null
+        sensitivity: null
+        time: null
+        flows:
+            set_categories: f.p
+    shape: [time, flows]
+
+Q_agg:
+    symbol: Q_agg
+    name: total demand by flow, aggregated
+    type: endogenous
+    coordinates_info:
+        scenarios: null
+        sensitivity: null
+        time: null
+        flows:
+            set_categories: f.p
+            aggregation_key: competition
+    shape: [flows, time]
+
+X:
+    symbol: X
+    name: total activity by technology
+    type: endogenous
+    coordinates_info:
+        scenarios: null
+        sensitivity: null
+        time: null
+        technologies:
+            set_categories: t.s
+    shape: [technologies, time]
+
+cap_n:
+    symbol: cap_n
+    name: new capacity
+    type: endogenous
+    coordinates_info:
+        scenarios: null
+        sensitivity: null
+        time: null
+        technologies:
+            set_categories: t.s
+        stock:
+            set_categories: k.n
+    shape: [time, technologies]
+
+cap_o:
+    symbol: cap_o
+    name: operative capacity
+    type: endogenous
+    coordinates_info:
+        scenarios: null
+        sensitivity: null
+        time: null
+        technologies:
+            set_categories: t.s
+        stock:
+            set_categories: k.o
+    shape: [time, technologies]
+
+C_op:
+    symbol: C_op
+    name: operational costs, total
+    type: endogenous
+    coordinates_info:
+        scenarios: null
+        sensitivity: null
+        time: null
+        technologies:
+            set_categories: t.s
+        flows:
+            set_categories: f.co
+    shape: [time, technologies]
+
+C_ca:
+    symbol: C_ca
+    name: capital costs, total
+    type: endogenous
+    coordinates_info:
+        scenarios: null
+        sensitivity: null
+        time: null
+        technologies:
+            set_categories: t.s
+        flows:
+            set_categories: f.ca
+    shape: [time, technologies]
